Refactor to create BlueOceanWebURLBuilder
